MessStaff model and service updated, contact and accountNum made Long from int

+++ MEAL_ON/src/main/java/com/mealon/MEAL_ON/service/MessStaffService.java
@@ -8,10 +8,10 @@ import com.mealon.MEAL_ON.model.MessStaff;
 public interface MessStaffService {
 	//Used only when MessAdmin is logged in
 
-	Boolean add(String name, int mess_id, int account_no, int contact, String address);
+	Boolean add(String name, int mess_id, Long account_no, Long contact, String address);
 	List<MessStaff> getAllMessStaff(int mess_id);
 	MessStaff get(int mess_id, String name);
-	Boolean update(String name, int mess_id, int account_no, int contact, String address);
+	Boolean update(String name, int mess_id, Long account_no, Long contact, String address);
 	Boolean delete(int mess_id, String name);
 	List<Integer> getStaffIdList(int mess_id);
